Tarjeta: ajustes en tests y descuento
Cambios de nombres de variables para más claridad. Se elimina un cast innecesario.

+++ ejercicio-tarjeta/src/main/java/ar/edu/untref/aydoo/Descuento.java
@@ -24,6 +24,6 @@ abstract class Descuento implements Beneficio {
     public int aplicar(final Compra compra) {
 
         int valorBruto = compra.calcularMontoBruto();
-        return (int) valorBruto - valorBruto * porcentaje / PORCIENTO;
+        return valorBruto - valorBruto * porcentaje / PORCIENTO;
     }
 }
